backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
faiss-cpu>=1.7.0
numpy>=1.24.0
pydantic>=2.0.0 